Rename and remove suggestName tests in favor of ProgrammaticPortsSpec
